\name{plates}
\alias{plates}
\docType{data}
\title{
Screening sample data
}
\description{
6 plates of high-throughput screening data in triplicates. Sample data is simulated. Sample names are random miRNAs.
}
\usage{data(plates)}
\format{
  A data frame with 1728 observations on the following 7 variables.
  \describe{
    \item{\code{Plate}}{a numeric vector}
    \item{\code{Replicate}}{a numeric vector}
    \item{\code{row}}{a numeric vector}
    \item{\code{column}}{a numeric vector}
    \item{\code{Sample}}{a factor with levels \code{} \code{Ctrl1} \code{Ctrl2} \code{CtrlKill} \code{GAPDH} \code{hsa-mir-1} \code{hsa-mir-1000} \code{hsa-mir-101} \code{hsa-mir-102} \code{hsa-mir-104} \code{hsa-mir-107} \code{hsa-mir-108} \code{hsa-mir-110} \code{hsa-mir-111} \code{hsa-mir-112} \code{hsa-mir-113} \code{hsa-mir-114} \code{hsa-mir-115} \code{hsa-mir-116} \code{hsa-mir-117} \code{hsa-mir-118} \code{hsa-mir-12} \code{hsa-mir-120} \code{hsa-mir-121} \code{hsa-mir-122} \code{hsa-mir-123} \code{hsa-mir-124} \code{hsa-mir-125} \code{hsa-mir-126} \code{hsa-mir-127} \code{hsa-mir-129} \code{hsa-mir-13} \code{hsa-mir-131} \code{hsa-mir-132} \code{hsa-mir-133} \code{hsa-mir-134} \code{hsa-mir-135} \code{hsa-mir-136} \code{hsa-mir-137} \code{hsa-mir-138} \code{hsa-mir-139} \code{hsa-mir-14} \code{hsa-mir-140} \code{hsa-mir-141} \code{hsa-mir-142} \code{hsa-mir-143} \code{hsa-mir-147} \code{hsa-mir-148} \code{hsa-mir-149} \code{hsa-mir-15} \code{hsa-mir-150} \code{hsa-mir-151} \code{hsa-mir-152} \code{hsa-mir-153} \code{hsa-mir-155} \code{hsa-mir-156} \code{hsa-mir-157} \code{hsa-mir-16} \code{hsa-mir-160} \code{hsa-mir-162} \code{hsa-mir-163} \code{hsa-mir-165} \code{hsa-mir-167} \code{hsa-mir-168} \code{hsa-mir-170} \code{hsa-mir-171} \code{hsa-mir-172} \code{hsa-mir-173} \code{hsa-mir-174} \code{hsa-mir-176} \code{hsa-mir-177} \code{hsa-mir-178} \code{hsa-mir-179} \code{hsa-mir-18} \code{hsa-mir-180} \code{hsa-mir-181} \code{hsa-mir-182} \code{hsa-mir-183} \code{hsa-mir-184} \code{hsa-mir-186} \code{hsa-mir-187} \code{hsa-mir-188} \code{hsa-mir-189} \code{hsa-mir-19} \code{hsa-mir-190} \code{hsa-mir-192} \code{hsa-mir-193} \code{hsa-mir-194} \code{hsa-mir-195} \code{hsa-mir-196} \code{hsa-mir-197} \code{hsa-mir-198} \code{hsa-mir-2} \code{hsa-mir-20} \code{hsa-mir-200} \code{hsa-mir-201} \code{hsa-mir-202} \code{hsa-mir-203} \code{hsa-mir-206} \code{hsa-mir-207} \code{hsa-mir-208} \code{hsa-mir-209} \code{hsa-mir-210} \code{hsa-mir-211} \code{hsa-mir-212} \code{hsa-mir-213} \code{hsa-mir-214} \code{hsa-mir-215} \code{hsa-mir-216} \code{hsa-mir-217} \code{hsa-mir-219} \code{hsa-mir-220} \code{hsa-mir-221} \code{hsa-mir-222} \code{hsa-mir-223} \code{hsa-mir-224} \code{hsa-mir-225} \code{hsa-mir-226} \code{hsa-mir-227} \code{hsa-mir-228} \code{hsa-mir-229} \code{hsa-mir-23} \code{hsa-mir-231} \code{hsa-mir-232} \code{hsa-mir-233} \code{hsa-mir-235} \code{hsa-mir-236} \code{hsa-mir-238} \code{hsa-mir-239} \code{hsa-mir-24} \code{hsa-mir-240} \code{hsa-mir-241} \code{hsa-mir-242} \code{hsa-mir-244} \code{hsa-mir-245} \code{hsa-mir-246} \code{hsa-mir-247} \code{hsa-mir-248} \code{hsa-mir-249} \code{hsa-mir-25} \code{hsa-mir-250} \code{hsa-mir-251} \code{hsa-mir-252} \code{hsa-mir-253} \code{hsa-mir-254} \code{hsa-mir-256} \code{hsa-mir-257} \code{hsa-mir-258} \code{hsa-mir-259} \code{hsa-mir-26} \code{hsa-mir-260} \code{hsa-mir-261} \code{hsa-mir-262} \code{hsa-mir-263} \code{hsa-mir-264} \code{hsa-mir-265} \code{hsa-mir-266} \code{hsa-mir-268} \code{hsa-mir-270} \code{hsa-mir-271} \code{hsa-mir-272} \code{hsa-mir-273} \code{hsa-mir-274} \code{hsa-mir-275} \code{hsa-mir-277} \code{hsa-mir-279} \code{hsa-mir-28} \code{hsa-mir-280} \code{hsa-mir-281} \code{hsa-mir-283} \code{hsa-mir-285} \code{hsa-mir-286} \code{hsa-mir-287} \code{hsa-mir-288} \code{hsa-mir-289} \code{hsa-mir-29} \code{hsa-mir-290} \code{hsa-mir-291} \code{hsa-mir-292} \code{hsa-mir-293} \code{hsa-mir-294} \code{hsa-mir-295} \code{hsa-mir-297} \code{hsa-mir-298} \code{hsa-mir-299} \code{hsa-mir-3} \code{hsa-mir-301} \code{hsa-mir-302} \code{hsa-mir-303} \code{hsa-mir-304} \code{hsa-mir-307} \code{hsa-mir-308} \code{hsa-mir-309} \code{hsa-mir-31} \code{hsa-mir-310} \code{hsa-mir-311} \code{hsa-mir-312} \code{hsa-mir-313} \code{hsa-mir-314} \code{hsa-mir-315} \code{hsa-mir-316} \code{hsa-mir-317} \code{hsa-mir-318} \code{hsa-mir-319} \code{hsa-mir-32} \code{hsa-mir-320} \code{hsa-mir-322} \code{hsa-mir-323} \code{hsa-mir-324} \code{hsa-mir-326} \code{hsa-mir-327} \code{hsa-mir-328} \code{hsa-mir-329} \code{hsa-mir-33} \code{hsa-mir-333} \code{hsa-mir-334} \code{hsa-mir-336} \code{hsa-mir-337} \code{hsa-mir-338} \code{hsa-mir-339} \code{hsa-mir-34} \code{hsa-mir-340} \code{hsa-mir-341} \code{hsa-mir-342} \code{hsa-mir-343} \code{hsa-mir-345} \code{hsa-mir-347} \code{hsa-mir-348} \code{hsa-mir-349} \code{hsa-mir-351} \code{hsa-mir-352} \code{hsa-mir-353} \code{hsa-mir-354} \code{hsa-mir-355} \code{hsa-mir-356} \code{hsa-mir-357} \code{hsa-mir-358} \code{hsa-mir-359} \code{hsa-mir-36} \code{hsa-mir-361} \code{hsa-mir-362} \code{hsa-mir-364} \code{hsa-mir-365} \code{hsa-mir-367} \code{hsa-mir-368} \code{hsa-mir-369} \code{hsa-mir-37} \code{hsa-mir-370} \code{hsa-mir-371} \code{hsa-mir-372} \code{hsa-mir-373} \code{hsa-mir-374} \code{hsa-mir-375} \code{hsa-mir-376} \code{hsa-mir-377} \code{hsa-mir-379} \code{hsa-mir-38} \code{hsa-mir-380} \code{hsa-mir-381} \code{hsa-mir-384} \code{hsa-mir-385} \code{hsa-mir-386} \code{hsa-mir-387} \code{hsa-mir-388} \code{hsa-mir-389} \code{hsa-mir-390} \code{hsa-mir-391} \code{hsa-mir-393} \code{hsa-mir-394} \code{hsa-mir-395} \code{hsa-mir-396} \code{hsa-mir-397} \code{hsa-mir-398} \code{hsa-mir-4} \code{hsa-mir-40} \code{hsa-mir-401} \code{hsa-mir-402} \code{hsa-mir-403} \code{hsa-mir-404} \code{hsa-mir-406} \code{hsa-mir-407} \code{hsa-mir-409} \code{hsa-mir-41} \code{hsa-mir-410} \code{hsa-mir-411} \code{hsa-mir-412} \code{hsa-mir-413} \code{hsa-mir-414} \code{hsa-mir-415} \code{hsa-mir-416} \code{hsa-mir-417} \code{hsa-mir-418} \code{hsa-mir-419} \code{hsa-mir-42} \code{hsa-mir-420} \code{hsa-mir-422} \code{hsa-mir-423} \code{hsa-mir-424} \code{hsa-mir-425} \code{hsa-mir-426} \code{hsa-mir-429} \code{hsa-mir-43} \code{hsa-mir-430} \code{hsa-mir-431} \code{hsa-mir-432} \code{hsa-mir-433} \code{hsa-mir-434} \code{hsa-mir-435} \code{hsa-mir-436} \code{hsa-mir-438} \code{hsa-mir-44} \code{hsa-mir-440} \code{hsa-mir-441} \code{hsa-mir-442} \code{hsa-mir-443} \code{hsa-mir-444} \code{hsa-mir-447} \code{hsa-mir-448} \code{hsa-mir-449} \code{hsa-mir-45} \code{hsa-mir-450} \code{hsa-mir-451} \code{hsa-mir-452} \code{hsa-mir-453} \code{hsa-mir-454} \code{hsa-mir-455} \code{hsa-mir-456} \code{hsa-mir-457} \code{hsa-mir-458} \code{hsa-mir-459} \code{hsa-mir-46} \code{hsa-mir-460} \code{hsa-mir-461} \code{hsa-mir-463} \code{hsa-mir-464} \code{hsa-mir-465} \code{hsa-mir-466} \code{hsa-mir-467} \code{hsa-mir-47} \code{hsa-mir-470} \code{hsa-mir-471} \code{hsa-mir-472} \code{hsa-mir-473} \code{hsa-mir-474} \code{hsa-mir-475} \code{hsa-mir-476} \code{hsa-mir-477} \code{hsa-mir-478} \code{hsa-mir-479} \code{hsa-mir-48} \code{hsa-mir-480} \code{hsa-mir-483} \code{hsa-mir-485} \code{hsa-mir-486} \code{hsa-mir-487} \code{hsa-mir-488} \code{hsa-mir-489} \code{hsa-mir-49} \code{hsa-mir-490} \code{hsa-mir-491} \code{hsa-mir-493} \code{hsa-mir-494} \code{hsa-mir-496} \code{hsa-mir-497} \code{hsa-mir-499} \code{hsa-mir-5} \code{hsa-mir-50} \code{hsa-mir-500} \code{hsa-mir-502} \code{hsa-mir-503} \code{hsa-mir-505} \code{hsa-mir-506} \code{hsa-mir-507} \code{hsa-mir-509} \code{hsa-mir-51} \code{hsa-mir-510} \code{hsa-mir-511} \code{hsa-mir-512} \code{hsa-mir-513} \code{hsa-mir-514} \code{hsa-mir-515} \code{hsa-mir-516} \code{hsa-mir-517} \code{hsa-mir-518} \code{hsa-mir-519} \code{hsa-mir-52} \code{hsa-mir-520} \code{hsa-mir-521} \code{hsa-mir-522} \code{hsa-mir-523} \code{hsa-mir-524} \code{hsa-mir-525} \code{hsa-mir-526} \code{hsa-mir-527} \code{hsa-mir-528} \code{hsa-mir-53} \code{hsa-mir-530} \code{hsa-mir-531} \code{hsa-mir-532} \code{hsa-mir-533} \code{hsa-mir-534} \code{hsa-mir-536} \code{hsa-mir-537} \code{hsa-mir-538} \code{hsa-mir-54} \code{hsa-mir-541} \code{hsa-mir-542} \code{hsa-mir-543} \code{hsa-mir-544} \code{hsa-mir-545} \code{hsa-mir-546} \code{hsa-mir-547} \code{hsa-mir-548} \code{hsa-mir-549} \code{hsa-mir-550} \code{hsa-mir-551} \code{hsa-mir-552} \code{hsa-mir-553} \code{hsa-mir-554} \code{hsa-mir-556} \code{hsa-mir-557} \code{hsa-mir-558} \code{hsa-mir-559} \code{hsa-mir-56} \code{hsa-mir-560} \code{hsa-mir-562} \code{hsa-mir-563} \code{hsa-mir-565} \code{hsa-mir-567} \code{hsa-mir-568} \code{hsa-mir-569} \code{hsa-mir-57} \code{hsa-mir-571} \code{hsa-mir-574} \code{hsa-mir-575} \code{hsa-mir-576} \code{hsa-mir-577} \code{hsa-mir-578} \code{hsa-mir-579} \code{hsa-mir-582} \code{hsa-mir-583} \code{hsa-mir-584} \code{hsa-mir-585} \code{hsa-mir-587} \code{hsa-mir-588} \code{hsa-mir-589} \code{hsa-mir-59} \code{hsa-mir-590} \code{hsa-mir-592} \code{hsa-mir-593} \code{hsa-mir-594} \code{hsa-mir-595} \code{hsa-mir-596} \code{hsa-mir-597} \code{hsa-mir-598} \code{hsa-mir-599} \code{hsa-mir-6} \code{hsa-mir-60} \code{hsa-mir-600} \code{hsa-mir-602} \code{hsa-mir-603} \code{hsa-mir-604} \code{hsa-mir-605} \code{hsa-mir-606} \code{hsa-mir-608} \code{hsa-mir-609} \code{hsa-mir-611} \code{hsa-mir-612} \code{hsa-mir-613} \code{hsa-mir-614} \code{hsa-mir-615} \code{hsa-mir-617} \code{hsa-mir-618} \code{hsa-mir-619} \code{hsa-mir-621} \code{hsa-mir-623} \code{hsa-mir-624} \code{hsa-mir-625} \code{hsa-mir-626} \code{hsa-mir-627} \code{hsa-mir-629} \code{hsa-mir-630} \code{hsa-mir-631} \code{hsa-mir-632} \code{hsa-mir-633} \code{hsa-mir-634} \code{hsa-mir-635} \code{hsa-mir-636} \code{hsa-mir-637} \code{hsa-mir-64} \code{hsa-mir-640} \code{hsa-mir-641} \code{hsa-mir-643} \code{hsa-mir-644} \code{hsa-mir-645} \code{hsa-mir-646} \code{hsa-mir-647} \code{hsa-mir-648} \code{hsa-mir-65} \code{hsa-mir-651} \code{hsa-mir-652} \code{hsa-mir-653} \code{hsa-mir-654} \code{hsa-mir-655} \code{hsa-mir-656} \code{hsa-mir-658} \code{hsa-mir-659} \code{hsa-mir-66} \code{hsa-mir-660} \code{hsa-mir-661} \code{hsa-mir-662} \code{hsa-mir-663} \code{hsa-mir-664} \code{hsa-mir-665} \code{hsa-mir-666} \code{hsa-mir-668} \code{hsa-mir-669} \code{hsa-mir-67} \code{hsa-mir-670} \code{hsa-mir-671} \code{hsa-mir-673} \code{hsa-mir-675} \code{hsa-mir-677} \code{hsa-mir-678} \code{hsa-mir-679} \code{hsa-mir-68} \code{hsa-mir-680} \code{hsa-mir-681} \code{hsa-mir-683} \code{hsa-mir-685} \code{hsa-mir-686} \code{hsa-mir-687} \code{hsa-mir-688} \code{hsa-mir-689} \code{hsa-mir-69} \code{hsa-mir-690} \code{hsa-mir-691} \code{hsa-mir-692} \code{hsa-mir-693} \code{hsa-mir-694} \code{hsa-mir-695} \code{hsa-mir-696} \code{hsa-mir-697} \code{hsa-mir-698} \code{hsa-mir-699} \code{hsa-mir-7} \code{hsa-mir-70} \code{hsa-mir-700} \code{hsa-mir-701} \code{hsa-mir-702} \code{hsa-mir-703} \code{hsa-mir-704} \code{hsa-mir-705} \code{hsa-mir-706} \code{hsa-mir-707} \code{hsa-mir-708} \code{hsa-mir-709} \code{hsa-mir-71} \code{hsa-mir-710} \code{hsa-mir-711} \code{hsa-mir-712} \code{hsa-mir-713} \code{hsa-mir-714} \code{hsa-mir-715} \code{hsa-mir-716} \code{hsa-mir-717} \code{hsa-mir-718} \code{hsa-mir-719} \code{hsa-mir-720} \code{hsa-mir-721} \code{hsa-mir-722} \code{hsa-mir-724} \code{hsa-mir-725} \code{hsa-mir-726} \code{hsa-mir-727} \code{hsa-mir-728} \code{hsa-mir-729} \code{hsa-mir-730} \code{hsa-mir-733} \code{hsa-mir-734} \code{hsa-mir-735} \code{hsa-mir-736} \code{hsa-mir-737} \code{hsa-mir-738} \code{hsa-mir-739} \code{hsa-mir-74} \code{hsa-mir-740} \code{hsa-mir-742} \code{hsa-mir-743} \code{hsa-mir-744} \code{hsa-mir-745} \code{hsa-mir-746} \code{hsa-mir-749} \code{hsa-mir-75} \code{hsa-mir-750} \code{hsa-mir-751} \code{hsa-mir-752} \code{hsa-mir-753} \code{hsa-mir-756} \code{hsa-mir-757} \code{hsa-mir-76} \code{hsa-mir-761} \code{hsa-mir-762} \code{hsa-mir-764} \code{hsa-mir-766} \code{hsa-mir-767} \code{hsa-mir-769} \code{hsa-mir-770} \code{hsa-mir-771} \code{hsa-mir-772} \code{hsa-mir-773} \code{hsa-mir-774} \code{hsa-mir-775} \code{hsa-mir-777} \code{hsa-mir-778} \code{hsa-mir-779} \code{hsa-mir-78} \code{hsa-mir-782} \code{hsa-mir-783} \code{hsa-mir-784} \code{hsa-mir-785} \code{hsa-mir-786} \code{hsa-mir-787} \code{hsa-mir-788} \code{hsa-mir-79} \code{hsa-mir-791} \code{hsa-mir-792} \code{hsa-mir-794} \code{hsa-mir-795} \code{hsa-mir-796} \code{hsa-mir-798} \code{hsa-mir-799} \code{hsa-mir-80} \code{hsa-mir-800} \code{hsa-mir-801} \code{hsa-mir-802} \code{hsa-mir-803} \code{hsa-mir-804} \code{hsa-mir-805} \code{hsa-mir-806} \code{hsa-mir-807} \code{hsa-mir-808} \code{hsa-mir-809} \code{hsa-mir-81} \code{hsa-mir-811} \code{hsa-mir-812} \code{hsa-mir-813} \code{hsa-mir-814} \code{hsa-mir-815} \code{hsa-mir-817} \code{hsa-mir-818} \code{hsa-mir-819} \code{hsa-mir-82} \code{hsa-mir-820} \code{hsa-mir-821} \code{hsa-mir-824} \code{hsa-mir-825} \code{hsa-mir-826} \code{hsa-mir-827} \code{hsa-mir-828} \code{hsa-mir-829} \code{hsa-mir-83} \code{hsa-mir-830} \code{hsa-mir-831} \code{hsa-mir-833} \code{hsa-mir-834} \code{hsa-mir-835} \code{hsa-mir-836} \code{hsa-mir-837} \code{hsa-mir-839} \code{hsa-mir-84} \code{hsa-mir-840} \code{hsa-mir-841} \code{hsa-mir-844} \code{hsa-mir-846} \code{hsa-mir-847} \code{hsa-mir-848} \code{hsa-mir-849} \code{hsa-mir-85} \code{hsa-mir-851} \code{hsa-mir-852} \code{hsa-mir-853} \code{hsa-mir-854} \code{hsa-mir-855} \code{hsa-mir-856} \code{hsa-mir-857} \code{hsa-mir-858} \code{hsa-mir-859} \code{hsa-mir-861} \code{hsa-mir-862} \code{hsa-mir-863} \code{hsa-mir-864} \code{hsa-mir-865} \code{hsa-mir-866} \code{hsa-mir-867} \code{hsa-mir-868} \code{hsa-mir-869} \code{hsa-mir-87} \code{hsa-mir-870} \code{hsa-mir-871} \code{hsa-mir-873} \code{hsa-mir-874} \code{hsa-mir-875} \code{hsa-mir-877} \code{hsa-mir-88} \code{hsa-mir-881} \code{hsa-mir-882} \code{hsa-mir-883} \code{hsa-mir-885} \code{hsa-mir-887} \code{hsa-mir-888} \code{hsa-mir-889} \code{hsa-mir-89} \code{hsa-mir-890} \code{hsa-mir-891} \code{hsa-mir-893} \code{hsa-mir-894} \code{hsa-mir-896} \code{hsa-mir-897} \code{hsa-mir-899} \code{hsa-mir-9} \code{hsa-mir-900} \code{hsa-mir-902} \code{hsa-mir-903} \code{hsa-mir-904} \code{hsa-mir-906} \code{hsa-mir-907} \code{hsa-mir-908} \code{hsa-mir-91} \code{hsa-mir-910} \code{hsa-mir-911} \code{hsa-mir-913} \code{hsa-mir-914} \code{hsa-mir-917} \code{hsa-mir-918} \code{hsa-mir-92} \code{hsa-mir-920} \code{hsa-mir-921} \code{hsa-mir-922} \code{hsa-mir-923} \code{hsa-mir-925} \code{hsa-mir-926} \code{hsa-mir-927} \code{hsa-mir-929} \code{hsa-mir-93} \code{hsa-mir-931} \code{hsa-mir-933} \code{hsa-mir-934} \code{hsa-mir-935} \code{hsa-mir-936} \code{hsa-mir-938} \code{hsa-mir-939} \code{hsa-mir-94} \code{hsa-mir-941} \code{hsa-mir-942} \code{hsa-mir-943} \code{hsa-mir-945} \code{hsa-mir-946} \code{hsa-mir-949} \code{hsa-mir-95} \code{hsa-mir-952} \code{hsa-mir-954} \code{hsa-mir-955} \code{hsa-mir-957} \code{hsa-mir-959} \code{hsa-mir-96} \code{hsa-mir-960} \code{hsa-mir-961} \code{hsa-mir-962} \code{hsa-mir-963} \code{hsa-mir-964} \code{hsa-mir-965} \code{hsa-mir-966} \code{hsa-mir-968} \code{hsa-mir-969} \code{hsa-mir-97} \code{hsa-mir-970} \code{hsa-mir-971} \code{hsa-mir-972} \code{hsa-mir-973} \code{hsa-mir-975} \code{hsa-mir-976} \code{hsa-mir-977} \code{hsa-mir-978} \code{hsa-mir-979} \code{hsa-mir-98} \code{hsa-mir-980} \code{hsa-mir-981} \code{hsa-mir-982} \code{hsa-mir-986} \code{hsa-mir-987} \code{hsa-mir-988} \code{hsa-mir-989} \code{hsa-mir-990} \code{hsa-mir-992} \code{hsa-mir-993} \code{hsa-mir-994} \code{hsa-mir-995} \code{hsa-mir-996} \code{hsa-mir-997} \code{hsa-mir-998} \code{hsa-mir-999}}
    \item{\code{Control}}{a factor with levels \code{GAPDH} \code{NEG} \code{POS}}
    \item{\code{CTB}}{a numeric vector}
  }
}
\details{
}
\source{
}
\references{
}
\examples{
data(plates)
}
\keyword{datasets}
